requirements.txt
streamlit==1.46.1
openai==1.70.0
anthropic==0.8.1
python-dotenv==1.1.0
pandas==2.2.3
numpy==1.26.4
matplotlib==3.8.0
pdfplumber==0.11.6
python-docx==1.2.0
watchdog==6.0.0
